// frontend/src/api/generated/models/projectBurnup.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectBurnupPoint } from './projectBurnupPoint';

export interface ProjectBurnup {
  project_id: number;
  project_name: string;
  started_on?: string | null;
  points: ProjectBurnupPoint[];
}